fix(file-storage): harden upload storage handling
Addresses PR review items for file metadata, validated size/content, repository-failure cleanup, adapter lifecycle, portable local key validation, MIME normalization, and bigint file sizes.

+++ packages/db/drizzle/0003_file_assets.sql
@@ -7,7 +7,7 @@ CREATE TABLE "file_assets" (
 	"storage_key" text NOT NULL,
 	"original_filename" text NOT NULL,
 	"content_type" text NOT NULL,
-	"size_bytes" integer NOT NULL,
+	"size_bytes" bigint NOT NULL,
 	"checksum_sha256" varchar(64) NOT NULL,
 	"status" varchar(16) DEFAULT 'uploaded' NOT NULL,
 	"created_at" timestamp with time zone DEFAULT now() NOT NULL,
